requirements.txt
azure-functions==1.20.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
